// src/openapi3/sdn-fabric/0.1.0/models/switch-model.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertyFabricId } from './additional-property-fabric-id';
import { SwitchModelAttributeList } from './switch-model-attribute-list';
import { SwitchModelPortList } from './switch-model-port-list';

/**
 * A model of switch (single)
 */
export interface SwitchModel {
  attribute?: SwitchModelAttributeList;

  /**
   * long description field
   */
  description?: string;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;

  /**
   * Pipeline configuration - dual or quad
   */
  pipeline?: 'unknown' | 'dual' | 'quad';
  port?: SwitchModelPortList;

  /**
   * unique identifier for the switch
   */
  'switch-model-id': string;

  [key: string]: 'unknown' | 'dual' | 'quad' | AdditionalPropertyFabricId | SwitchModelAttributeList | SwitchModelPortList | string | undefined;
}