// src/openapi3/sdn-fabric/0.1.0/models/switch-model-port.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { SwitchModelPortSpeeds } from './switch-model-port-speeds';

/**
 * A port in a switch (single)
 */
export interface SwitchModelPort {

  /**
   * identifier of the cage - physical port on switch
   */
  'cage-number': number;

  /**
   * long description field
   */
  description?: string;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;

  /**
   * A splitter can divide the port in to channels.
   * The default value 0 indicates the port is not channelizable
   */
  'max-channel'?: number;
  speeds?: SwitchModelPortSpeeds;
}